backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
mediapipe==0.10.14
Pillow==10.3.0
numpy==1.26.4
websockets==12.0
pytest==8.2.0
httpx==0.27.0